tcgbios: Add physical presence interface support
Add support for TPM 1.2 and TPM 2 physical presence interface
(PPI). A shared memory structure is located at 0xffff 0000 -
0xffff 00ff that SeaBIOS initializes unless it has already
been intialized and then searches for a code it is supposed
to act upon. A code typically requires that one or more TPM
commands are being sent.

The underlying spec can be accessed from this page here:

https://trustedcomputinggroup.org/tcg-physical-presence-interface-specification/

Version 1.20 is implemented.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>